Committer: Costin Badescu <costin@riemann.localdomain>  On branch master  Changes to be committed: 	modified:   src/core/OMakefile 	modified:   src/core/lexer.ml 	new file:   src/core/logic.ml 	new file:   src/core/logic.mli 	modified:   src/core/parser.ml 	modified:   src/core/reconstruct.ml 	modified:   src/core/syntax.ml 	modified:   src/core/syntax.mli 	modified:   src/core/trail.ml 	modified:   src/core/trail.mli 	modified:   src/main/main.ml
Committed to Beluga:

1. Cleaned up implementation of trailing. (trail.ml[i])
2. First commit of the code for the logic programming engine. By default,
the engine is turned off. The '-logic' switch activates it. The
implementation is still incomplete. Simple queries should work well. The
generated proof terms are not type checked yet. They are only printed to
STDOUT. To interact with the LP engine, enter %query followed by a correct
LF type of the logical statement to be proved. Warning: depending on the query, the proof-search may not terminate.