Use portable ln-or-cp rather than symlinks and fix preexisting bin/ issue.
omake seems to have a bug when testing for the existence of
directories. Work around it by removing the existence test, which was
redundant anyway, given that we can just pass -p to mkdir.

I also took the opportunity to remove a bit of syntactic noise in the
root OMakefile.