Remove multiple session support.
This allowed one to pass multiple arguments to Beluga, and have it
type check each .bel or .cfg one by one, without forking a new process
for each. This is a misfeature, however, because it made testing
difficult:

 - global state was not guaranteed to be completely and correctly
   reset, potentially making checking of one file fail, but only when
   it is checked after some other file. If instead we fork a new
   process each time, we are guaranteed to start each run from a clean
   slate.

 - it made the Beluga process long lived, which in turn caused
   problems with timeouts, such as the one Costin added to logic.ml to
   catch infinite loops.

main.ml is now, as a result of dropping this feature, quite a bit
shorter.

Note that this does not affect the ability to check multiple related
files in one go, by supplying one .cfg file as argument.