Add support to TEST for comparing output with .out files.
When TEST runs a test case foo.bel, it records the output of Beluga.
If foo.bel.out exists, then TEST reports SUCCESS so long as the output
matches the content of that file. If it does not, then it displays a
(colorized!) diff of the two. If foo.bel.out does not exist, then it
reports success, or failure, as before, depending Beluga's exit code.

This makes it possible to test for expected failure.