Introduce .dir-locals.el.
This file sets project local variables in Emacs. For example, mode
lines of form

(* -*- coding: utf-8; indent-tabs-mode: nil; -*- *)

at the beginning of .ml* files are no longer necessary, because the
coding and indent-tabs-mode variables can be set once and for all for
every file in the Beluga project in .dir-locals.el. Note that the
settings in .dir-locals.el normally do not affect other buffers from
other directories. This feature was introduced with Emacs 23.

Currently sets coding, indent-tabs-mode and compile-command. The
latter is useful to override the default "make -k" command to use
omake instead.