Deal with pattern variables in opsem.
Matching in the presence of pattern variables means extending the
ordinary substitution eta with new values as we traverse the spine of
a pattern. This is in contrast with how we deal with meta-variables in
patterns, because for those we basically start with a
meta-substitution of the right size and get unification to alter that
meta-substitution by side effect.

This asymmetry is kind of unsatisfactory, so more work is needed.