Remove injecting values into expressions.
This was only needed because Opsem.eval would take expressions to
expressions, so the value obtained by Opsem.eval needed to be injected
into an expression.

But having such an injection was bad, because it meant that we had
two ways of representing ground terms, eg (Boolean true) and
(Value (BoolValue true)), both of the same type. This was evident in
the Opsem.eval, which reinterpreted some select values into
expressions that did not involve the Value constructor. Because this
reinterpratation was limited to only the head of the value, it did not
reinterpret subvalues of composite values, neither was that possible,
because it was only possible to embed values within expressions, not
the other way round.

Removing this injection has the side effect of forcing us to write a
more precise type for Opsem.eval, which is now:

eval : Comp.exp_chk -> Comp.value

But with this new type for 'eval', we are in turn forced to make the
bodies of global constants values, not expressions, because when a
constant is defined using a toplevel let we call eval on its body,
yielding a value that now cannot be injected into an expression.
Therefore I have also changed the store of constants so that constants
are always values, as does indeed make sense under a call-by-value
semantics, rather than expressions.