Massive code quality cleanup to make codebase compile under Ocaml 4.00.
This new version of the compiler is much better at finding unused
definitions in the code and needlessly recursive lets. The majority of
the changes in this commit are to make many recursive lets non
recursive, and the removal of functions that were unused throughout
the codebase (ie dead code).