lexer: correct token spans
Beluga would use the wrong spans for tokens. As an example, consider the
file

```
rec foo : Bool = ttrue
```

The token `rec` would be assigned the span 0-0, since the initial value
for the source location has the start and stop offset 0. By calling
`lex_token` before dereferencing the location, we ensure that a call to
`Loc.shift` occurs, thus adjusting the location to `0-3` as expected.