lexer: correct token spans and improve newline handling
Beluga would use the wrong spans for tokens. As an example, consider the
file

```
rec foo : Bool = ttrue
```

The token `rec` would be assigned the span 0-0, since the initial value
for the source location has the start and stop offset 0. By calling
`lex_token` before dereferencing the location, we ensure that a call to
`Loc.shift` occurs, thus adjusting the location to `0-3` as expected.

As for newlines, we need to do a `shift` to cover the location of the
newline character *as well as* an `advance_lines` to bump up the line
number. Otherwise, location offsets reported by Beluga are off by the
number of lines. (e.g. If Beluga is reporting the offset of a token on
line 55, then that offset will fall 55 characters short of the tokens
true position in the.)

This commit refactors the shifting and advancing logic somewhat to make
it less repetitive (and error-prone!)