add %:lookuphole interactive command
This command takes a single hole identifier and determines what its hole
number is. This is essential for the Emacs code, which used to use hole
numbers as a way of determining which overlays to affect. But now, hole
numbers are not the only way to identify a hole, so we need to ask
Beluga to translate hole names into numbers for us.

I foresee this command being removed later, in favour of storing more
information inside a hole data structure on the Emacs end. This way,
once Emacs knows everything about every hole (essentially, we should
serialize the `hole` type and send it to Emacs) Emacs can draw the hole
information in whatever way is most convenient for the user. What's more
is that this gives us flexibility to design interactive modes with other
editors.