Integrate lexer testing with test harness
This obviates the need for the scripts lex_check.sh and lex_check_all.sh
that were previously introduced.
Now, when running the test harness, the program lex_dump will be used to
compute where Beluga thinks the input ends and wc will be used to
compute where the file *really* ends. If these counts are mismatched,
then the message LEXER FAILURE will appear (in cyan) before the semantic
test status (DONE, FAIL, TIMEOUT, etc.).
The total count of lexer failures is tallied and displayed in the test
summary at the end.