don't stage generated holes
If we stage new holes that are generated (by splitting or by intro),
then we need to eventually commit those holes and delete the holes
they replace. Otherwise, these holes will be become erroneously
committed on the next file load.
Instead, we can simply avoid staging the holes at all. In this case,
the new holes simply exist to be pretty-printed. Indeed, the split and
intro commands simply generate a small chunk of Beluga internal syntax
and pretty-print it; they don't _modify_ the AST of the currently
loaded program.

To pull this off, we have to make an assumption about these so-called
"fresh holes". Recall that the internal syntax Hole constructor has
three fields: its location, its name, and a thunk that will compute
the number of the hole.
We can't give a legitimate thunk to compute the hole number, since we
never stage (and hence never commit) the fresh holes. This isn't a
problem though: that thunk will never be evaluated. So we use a thunk
that simply raises an exception.

This fixes #46 which caused bogus hole locations to appear for the new
holes. The specific sequence of events that caused #46 is the
following:
1. Load a file with one hole.
2. Split on that hole.
   Now the problem happens: splitting will generate new holes, which
   will be staged but not committed. Hence, they will just stay in the
   staged holes array until the next %:load command.
3. Load a file. (Doesn't necessarily need to be the same one.)
   When loading a file, only the committed holes are cleared, so the
   staged holes array from before has persisted.
   Holes in the new file are staged during parsing, and once parsing
   completes, all holes are committed, including the fresh holes from
   before. Those holes now have bogus locations since hole locations
   are the only way to uniquely identify holes.

As of this commit, we no longer stage the holes, and we don't do this
bogus hole location trick to identify the holes anymore. In fact, it
wasn't necessary to identify the fresh holes at all, so we use
exceptions.