remove staged holes
Instead, holes are directly added to the array of holes during
typechecking. The hole array does not support deleting holes as of
yet, since this would cause a downward shift of all hole indices in
later positions, thus corrupting hole identifiers used elsewhere.

I'm starting to think that the hole array should in fact be immutable:
new holes should not be added to it or removed from it during the
lifetime of a loaded file.