fix #68
When indexing types, (translating external syntax into approximate
syntax,) we presently forbid holes from appearing in types.

Generally, types must be written by the user (for `rec` declarations),
in which case putting a hole in the type doesn't make sense. (The body
of a `rec` declaration is a _checkable_ expression, not a
_synthesizable_ expression.)
However, for `let` declarations, the body must have a synthesizable
type, so it intuitively should be possible to use a hole for the type
in order to get Beluga to tell the user what the type is.