refactor LF declaraction plicity check
It turns out that multiple places had essentially the exact same
function. In particular, a function `is_inferred` was defined in
`interactive.ml` and a function `isExplicit` was defined in
`lfholes.ml`. Behaviourally, one simply computes the negation of the
other.

Although the version in `interactive.ml` had been patched when
resolving #82 in ec2c3cf74f, the patch wasn't propagated to
`lfholes.ml` because I didn't know that the function had been
duplicated.

In this commit, we remove `isExplicit` from `lfholes.ml` and
`is_inferred` from `interactive.ml` and instead define
`LF.is_explicit` in `synint.ml` as a one source of truth for the
implementation of this function.

Fixes #74.
Related to #77.