combine Lfholes with Holes
Lfholes was essentially a copy of Holes, so I simply generalized Holes
to treat both computational holes and LF holes.
This in turn eliminated the duplication of certain commands in the
interactive mode which had to be specific about what type of hole they
operated on. Now there is simply "printhole" instead of both
"printhole" and "printhole-lf".
I removed the "numholes" and "numholes-lf" commands to instead make
"countholes" print the total number of holes; I adjusted the Replay
tests to account for this.
LF holes now support names.