logic: add meta-context and projection search
meta-context search
-------------------

This improves the logic programming engine so that it analyzes the
meta-context for solutions as well. Currently the --query pragma does
not support meta-contexts, but the %:solve-lf-hole interactive command
will use the meta-context of the hole to augment the search.

To make this work some slight fiddling with the de bruijn indices is
necessary.
Basically, when we pull something out of the meta-context to use it,
we need to weaken it according to its position in the meta-context.
For instance, consider a meta-context with four entries "M, A, B, D".
If we want to see if D will match the goal type, we need to weaken D
by one, since the goal type exists in the meta-context with four
entries whereas D exists in the meta-context depending only on the
three preceding terms!
So the introduced `matchDelta` function that traverses the
meta-context is augmented with a counter used to appropriately weaken
the context entry before trying to unify it with the goal.

projection search
-----------------

Sigma-types, arising from parameter variables of the form `#p`, now
have all their projections tried. This is a standard proof search
strategy, namely to break down product types as assumptions.

This is somewhat tricky because these are dependent products, so when
we traverse a sigma type and try its nth projection, we form its type
by substituting it all terms up to n-1.

The matchSigma function implements this logic by carrying with it an
LF substitution that it applies to the type of the nth projection when
unifying it with the goal.