coverage: add LF normal term as starting point for splitting
If the case expression is on an LF normal term, then we can use that
term as the starting point for splitting. Consequently, the coverage
checker will detect more impossible cases.

The feature is disabled because it causes some tests to fail. To
enable it, change from "None" to "Some (l, mT)" in check.ml; there is
a comment in the file suggesting this.