hook up Harpoon interactive mode
The Harpoon interactive mode is started from the basic interactive
mode by using the %:prove command and entering the statement of a
theorem to prove. The Harpoon interactive mode's prompt is a lambda.

Currently, only the `intros` tactic is implemented, and will introduce
all the assumptions.
The `show-proof` command will display the current proof script.