store: add conversions from contexts
Harpoon needs to elaborate terms, so in particular it needs to index
them. Term indexing threads along lists of names so that it knows what
index to use when a name is encountered. What we track in Harpoon is
contexts, so we need a way to recover a name list from a context in
order to do elaboration.