synext/parser: add harpoon commands
Because Harpoon commands may contain embedded Beluga expressions, I
define the syntax of the interactive Harpoon commands as external
syntax and add parsing rules for them in the core Beluga parser.
This makes it easy to parse terms.

However, it turns out that we can't make the harpoon commands parse
simply by their names, since for some reason this creates an ambiguity
in the parser. For example, if we had used the terminal "split"
instead of "%:split" in the parser, then ordinary Beluga code trying
to define a "split" type will fail to parse. It is unclear to me why
this is happening, so as a workaround for now, Harpoon commands will
be lexically distinct from ordinary Beluga identifiers.