harpoon: refactor; use delayed msub
This is the way things are done everywhere else. Rather than eagerly
normalize types, we carry around `tclo` objects, which contains an
open type and a grounding substitution.

We also isolate the command processing logic into its own function,
called after the command is parsed.

I adjusted Claim to take optionally a term to justify the claim.

I adjusted the pretty-printer to deal with the tclo goals in proofs as
well as the augmented Claim constructor.
The pretty-printers also all carry around contexts now, so that
rendering the embedded terms is possible.