store: relax contextual variable name comparison
CVar.index_of_name was using structural equality (`=` operator) to
compare `name` records. In ordinary Beluga operation, this is fine
because it only happens during indexing, at which point there are no
system-generated names in the system, so only the string parts of the
names contain something. However, this is unacceptable in Harpoon
because I need to elaborate terms typed by the user in the context of
an already loaded file. Concretely, I'm already in a context
(Delta;Gamma) which I erase to a store in order to index new types;
the names in that store will hence not be structurally equal to the
names generated by the parser, despite having identical string
representations.

I add the function `equals` to the Id module to compare string
representations of names, and use this as the equality in Store.CVar.
Without this, type synthesis would fail for terms such as [ |- X1 ]
even when a variable (seemingly) named X1 was in the context!

I will probably need to relax the equality of other name comparisons,
but I will leave that for later.