harpoon: have tactics manipulate subgoals
There was a spread of responsibilities between the main command
processing loop and the tactics in regards to managing the list of
subgoals: tactics were provided a callback to register new subgoals
that they created, and they always had the responsibility of solving
the current subgoal, since the command loop would call
`remove_current_subgoal` after invoking the tactic.

If this is confusing to you, don't worry. It's confusing. And that's
why we replaced it.

Instead, tactics are wholly responsible for managing the subgoal list:
tactics are expected to register any new subgoals they are added, and
are expected to remove the current subgoal if they solved it.

This is nice because it moves some work out of the main loop.

Structurally, we use a module for representing the "tactic context",
which currently has only two capabilities: `add_subgoal` and
`remove_current_subgoal`. The `Tactic` module is now parameterized by
a tactic context (so it is a functor) and we instantiate it in the
main loop.