add `invert` tactic & start working out local context tracking
The `local_context` field tracks *new* variables introduced in the
current scope. This is then used when pretty-printing so we don't
reiterate the entire context in every scope.
It isn't perfect yet, because splitting can change the *types* of
variables that were previously in scope, and it isn't clear what to do
about that yet. Should variables whose types changes be re-listed in
the local context?
In any case, I haven't worked out how to figure out which variables'
types changed.