holes: use a hashtable instead of a list
We may want to sort holes by location before printing them now that
the storage order is arbitrary. Using a hashtable has the benefit that
deleting entries won't affect the keys using for other entries. The
previous list-based implementation was somewhat unstable in that
regard, and essentially it assumed that you wouldn't be deleting and
adding holes a lot. However, deleting and adding holes will be how
Harpoon interacts with LF proof search, so we need to make sure that
holes are well-behaved in that sense.