holes: add snapshots
This way we can find holes that were added between reconstructions of
expressions. Harpoon can use this to detect if a hole is present in an
expression typed by the user. Then we can use the logic programming
engine to solve LF holes automatically.

Eventually we will need to add a "solution" mutable field to holes so
that the solutions found by proof search can be filled in. The
pretty-printer and typechecker will need to be made aware of this new
field so they can pretend the hole doesn't exist and just check the
solution.