holes: add solutions
Now holes come bundled with a _solution_, which is a mutable field
that can be filled in later.
This gives the following flow for using LF proof search in Harpoon:
- The user types a command involving an expression with a hole,
  e.g. `solve [|- ?]`
- This is parsed into external syntax, indexed into approximate
  syntax, and reconstructed into internal syntax.
  During reconstruction, holes in synthesizable positions are staged.
- The reconstructed expression is typechecked, which stages holes in
  checkable positions.
- Using the snapshots system, we are able to enumerate the holes in
  the expression.
- We solve each hole in the expression using LF proof search, and fill
  in the solution field.
- We typecheck the whole expression again, now that solutions have
  been filled in. The typechecker will now check that the solutions
  are correct.
- We check that there are no holes remaining in the expression.

It isn't totally clear how this works with dependent types yet.
For example, if we have a function of two arguments, with the second's
type depending on the first, and the first argument is given as a hole
to solve, there's presumably some kind of MMVar stuff going on behind
the scenes that this solution system is not interacting with.

The legacy commands (in command.ml) that deal with holes now require
unsolved holes at all times.