harpoon: always double-check reconstructed expressions
This is crucial for catching holes that are staged by checking! Holes
in synthesizable positions are staged during reconstruction, but holes
in checkable positions are only staged during typechecking.

It's unclear whether it *has* to be this way.
If I had it my way, reconstruction would stage *all* holes, and store
the allocated hole id inside the internal syntax hole. Then checking
can lookup the hole from the global holes table, check if it's solved,
and check the solution if any.