store: make storing evaluated programs optional
This is crucial for Harpoon because we would like to add the theorem
to prove to the store so that it can be found when we index
expressions entered during the Harpoon session (for IH checking).
Previously, we worked around this by adding the theorem to cG when
elaborating `by ih`, but this is quite messy as it messes up the
indices. Instead, we can add the theorem to the store so that indexing
picks it up the same way it does is ordinary Beluga programs.