check: proper error for unexpected free variable
This would produce an internal error in Harpoon:

> by ih foo [ |- c v ] as ih

where `v` is some free variable.
It gets disambiguated to a free metavariable, which had no case in
`inferHead` in lfcheck.ml, thus producing a Match_failure.

Now we instead correctly raise LeftoverFV plus a hint that the
variable may be misspelled. (This IME is the most common cause of
leftover free variables.)