harpoon: add `unboxed` flag for `by`
When using a `by` command, one can now write `unboxed` at the end in
order to immediately unbox the result of the invocation and add it to
Delta, thus avoiding the need to separately unbox the result of the
invocation. This also avoids adding dummy entries to Gamma jsut to
unbox them one step later.

The full syntax is:
  by C t as x [unboxed|boxed]

where
 - C is an invocation kind, either `lemma` or `ih`
 - t is a synthesizable term (an application)
 - x is a (meta)variable name