harpoon: add `rename` command
Allows renaming assumptions in Gamma or Delta. Currently, the level
must be specified in the command, but in principle this can be
inferred, as long as there is no ambiguity.

The renaming is *local* and will only affect the current branch of the
proof, so it's best to rename things as soon as possible, broadly
speaking.

No sanity checking of the names is performed, so if you rename
something so that it has the same name as something else, tough luck.
In general its possible to use the rename command several more times
to resolve this kind of situation. Even then it's not the end of the
world if there are duplicate names, as long as you don't want to try
referring to one that's "further away".