remove EmptyBrach and PatEmpty; disallow `in` with `impossible`
Instead, we add `impossible` to the (approximate syntax and to the)
internal syntax, and use a function Coverage.is_impossible to decide
when a type is impossible, by enumerating coverage goals for that type
and checking that there are none.
This makes the syntax of patterns simpler (empty patterns are no
longer allowed anywhere) and it removes some invariants from the code
(e.g. an empty pattern must not have a RHS).

It turns out that `in` with impossible was completely broken: in the
absence of an `in`, the LF type was assumed to be in a closed context,
so cases for parameter variables were not generated. This meant that
using the LF empty pattern {} was strictly more powerful than
impossible. Now `in` is removed, and the context in which a type is
checked to be impossible exactly the type of the given term that
witnesses the uninhabited type.

This closes #109.
This closes #107 as well, but it isn't totally clear why.
This closes #73 probably because `impossible` is treated more carefully.