replace Ann with AnnBox
Beluga does not generally allow arbitrary type annotations, but it
does crucially use them for handling boxes in synthesizable positions,
since in general boxes are only allowed in checkable positions.

So instead of using a general Ann constructor for a type annotation,
and externally requiring the subterm to always be a box (using
inexhaustive pattern matching), I have replaced Ann with AnnBox, which
necessarily contains a meta_obj.