harpoon: fix dependent pattern matching
In the process, I needed to rewrite Reconstruct.synPatRefine to be
more sane.

synPatRefine is hoisted out of the big mutual and-block, since it
doesn't call any of those functions.

Rename `PatAnn` exception `UnsupportedTypeAnnotation`, which is more
descriptive.

It now has more general preconditions (previously its preconditions
were tightly coupled to the way type reconstruction did things --
perhaps it used to be inline within elBranch and was hoisted out more
or less as is).
The invariants of synPatRefine are now clearly documented.
Harpoon's `split` command now uses synPatRefine in order to correctly
support dependent pattern matching.

Closes #132