coverage: make it work correctly with substitution variables
Coverage would always succeed when analyzing computation types indexed
by substitution variables, since it would erroneously believe that
these types are uninhabited. This is due to `genPattSpine` only
considering the PiBox case for a ClTyp MTyp, and not a ClTyp STyp. I
added the case for ClTyp STyp, and adjusted the implementation of
matching to account for it.

This exposed that one of our tests was not covering, so this test has
been fixed.

This impacts Harpoon crucially because we use genPatCGoals from
coverage to generate the cases to consider when doing a
split. However, genPatCGoals used to generate 0 cases for computation
types indexed by substitution variables.
In core Beluga, this just produced false positive coverage checks, but
in Harpoon it broke splitting.