make sigils part of a variable's name
Notably, this allows you to have a proper MVar named S as well as a
substitution named $S, which was previously impossible because SVars
(and PVars) had their sigils stripped during lexing.

A somewhat strange consequence of this is that when svars (or pvars)
are introduced using mlam, you need to name them with their sigil.