harpoon: add multi-session support
Closes #153

This patch adds a 'hidden' flag to Store.Cid.Comp.entry. Hidden
entries cannot be found during indexing. When a session is suspended
(so you can go prove a lemma) all its theorems become hidden.

The multi-session support could be nicer, because as it stands you
must prove the lemma before you are allowed to use it. It would be
nice to able to assume that the lemma holds, use it to prove the
theorem, and then go back to prove the lemma. This is like a
theorem-level 'suffices' command.