harpoon: refactor prover; add command sequences
Closes #160

Users can write multiple commands in one line separated by semicolons.
I added a restriction that the commands must all apply to the same
theorem. That is, if your sequence completes on any step but the last,
you get a warning that the remaining commands were skipped.

If a command errors, the remaining commands are also skipped.