harpoon: induction fixes
- Numeric total orders are elaborated in such a way as to account for
  abstraction's surprising behaviour regarding context
  quantifications.
  See issue #173
- Total.annotate is called during theorem configuration (as late as
  possible).
  Previously, I annotated the type as early as possible and later
  stripped the annotations when necessary, but it turns out that
  Total.strip is NOT an inverse of Total.annotate. In fact,
  Total.annotate is simply not invertible in the current
  implementation because of the LF.Inductive annotation not saying
  whether the variable is explicit or implicit.

I'm not proud of the second point, because calling Total.annotate
later delays any error until when the user has configured *all* the
theorems. Instead, one could call Total.annotate once, early on, to
check that the order is compatible with the type, but throw away the
result until later when we recalculate it. Alternatively (and this
would be more work) one could pass around both the type and its
annotated form until theorem configuration. Then the annotated form
can be used.