use more precise total decs
Total.dec was moved previously into Syntax.Int.Comp.
Now we change it up so that we can represent different kinds of
totality checking that we have. We consider partial functions, trusted
functions, and non-recursive functions to be variant forms of totality
checking, with the main form of checking represented by the `inductive
tag.

Previously, it was impossible to later determine whether a function
was nonrecursive, partial, or whatever. All we knew was that it
"didn't require checking" and other information was used as a proxy
for figuring out whether it was a total function or not.
In particular, this simplifies the check for whether a total function
is calling a partial function.