reconstruct: adjust synPatRefine
Now the pattern argument to synPatRefine is less required.
Specifically it isn't necessary at all anymore when we know for sure
that we're generating a refinement for a computational split (on an
inductive/stratified type).
Instead of being a proper argument to synPatRefine, it is passed via
the case_type object and bundled together with the scrutinee, since
the purpose of the pattern is precisely unification against the
scrutinee in the case of proper dependent pattern matching.