coverage: slightly refactor genObj
We would like to use genObj in Harpoon elaboration to generate a
pattern for use in synPatRefine. However, genObj returns a triple
containing a coverage goal. This would require partial pattern
matching in Harpoon. It's best to avoid partial pattern matching when
possible. The fix here is to just make genObj return a nested tuple
structure that coverage can internally package back up into a CovGoal
when it must do so.