check: fix unroll and make it return an msub, too
This msub is a weakening, to apply to things like cIH.
As for the fix, `unroll` would incorrectly extend cG. Now this
extension is done is stages. First we build up a new context cG',
computed from the TypArr parts of tau, such that cD' |- cG'. Then we
apply the weakening to the input context cG, and append cG'.
See the code comments for a more detailed account of this approach.