check: store subgoal with type from check
During checking, when we encounter an incomplete proof, holding an
uninstantiated subgoal, we need to add it to the goal list in
holes.ml. However, we can't add it verbatim, because the goal and
context associated initially assigned to that subgoal were generated
by type reconstruction, which does not do totality checking.
Therefore, the initially associated goal and context are
missing totality checking metadata (LF.Inductive / Comp.TypInd). This
will cause the recovered Harpoon session to be broken, such that it
always believes recursive calls aren't structurally smaller.