harpoon: improve subgoal paths (& reconstruction); fix `reset`
Subgoal paths used to merely be lists of strings, but now there are a
nice data structure. I realize that that the path might be better
represented as a list of components, but the current structure works
fine.

Subgoal paths are correctly reconstructed, and are generated in the
same way by Harpoon interactively.

reset now uses the equality for subgoal paths, defined in Whnf.