harpoon: use subgoal paths to lookup the subgoal to remove
If we try to ordinary equality as before, we get an error because
subgoal paths are higher-order.
I tried to implement a convertibility between subgoals, but it doesn't
make sense unless we can check convertibility of contexts in a
meaningful way. Otherwise we might try to convert the goal types that
don't live in the same Delta.
However, convertibility of subgoal paths works fine, and subgoal paths
uniquely determine a subgoal.