prettyint AnnBox metaobj using its type for better output
In particular, when printing an AnnBox, we can use the type in order
to obtain the full LF context Psi when printing the term. The upshot
is that we can generate the type annotations necessary on bound
variables to guarantee that type reconstruction succeeds.

See #182, #183 for potential solutions that don't require type
annotations.