index: case labels use vars from inside the branch
This is because context case labels identify the schema element by
type. And this is tricky, because these types are not *closed*! They
can refer to existentially-bound variables. So in order to interpret
these types correctly, we need to understand them in the branch
context, after the refinement has occurred. The result is that we
index the body of the branch before its label.

Note from the future:
This turns out to be a massively bad idea, and so I eventually changed
context case labels to identify the schema element by number.
An additional reason to not use a type-based identifier for the schema
element is that in general, it can be ambiguous. What are the case
labels for a schema like `unit + unit`?