coverage: refactor genCtx; add order warning
Rather than put the empty-context case at the end, we put it at the
beginning. This enables us to use List.map to generate the schema
element cases, and produces a case ordering that matches the 1-based
schema element indices.

I added a very scary comment to genCtx warning that Harpoon crucially
relies on the ordering produced by this function.