reconstruct: allow matching on Sigma-type expressions
In particular this enables matching on a PVar in a nonempty context if
the schema involves a Sigma.

This patch also enables partially-applied LF constructors to have
function types correctly synthesized for them. This is not apparently
useful in Beluga since `x : [|- term -> term]` can't be matched on to
obtain a metavariable `X : (|- term -> term)`, but in Harpoon it's
possible.