coverage: refactor; use contextual name generation
Name generation in coverage no longer uses a global variable-based
name generator, and has instead switched completely to contextual name
generation for all its variables.

I have slightly refactored some functions to over unnecessarily
injecting their returned value into the covgoal type, only to be
inverted later. In particular, genPatCGoals was guaranteed to return
something wrapped with CovPatt, so Harpoon had to use partial pattern
matching to retrieve the real return value. Now it just returns a
tuple. To make these massive tuple types more manageable, the `inside`
type constructor introduced previously is used.

genPatCGoals used to take two gctx parameters as input. It seems this
is for historical reasons, as there were no call sites in the codebase
that ever used something other than Empty for the second context. The
first gctx parameter is the ambient gctx, which genPatCGoals would
extend with new computational variabels in case of a split on a
computation type. This is unnecessary, since genPatCGoals's caller can
just append the ambient gctx to the gctx generated by genPatCGoals.
So the type of genPatCGoals is now much simplified.

There is a very large seeming change in tactic.ml but this is just the
result of unindenting code due to eliminating cases to produce
violations if genPatCGoals didn't return a CovPatt previously.
Changes in modules outside coverage.ml are simply to accommodate the
change in the types of genPatt and genPatCGoals.