don't use polymorphic equality
All instances of the polymorphic equality function are now replaced
with more specific equality functions that are safer. Primarily, this
is to eliminate comparisons of Id.name values using polymorphic
equality, which is totally broken in the presence of location
information attached to names. This is because the polymorphic
equality will also compare the locations, which is not desired.

Things that indirectly use polymorphic equality, like Hashtbl and
List.mem are also replaced.