abstract: use contextual name generation
Abstraction produces a meta-context that is appended to the ambient
meta-context by type reconstruction when analyzing a
pattern. Name-based lookups are performed by Apxnorm into the
resulting meta-context in order to determine the offset of free
variables inside a branch.

However, nothing was stopping abstraction from producing duplicate
names that would throw off the lookups. These duplicate names were
causing very strange type-checking failures in a test, because a
variable in that test wound up being assigned an offset to an
instantiated MMVar instead of to a binder in a pattern. This happened
precisely because both had the same name, and name comparisons are now
done with the equals function, that looks at textual representations.
The reason this test worked before is that structural equality also
considered the `counter` field in the name, which happens to be
incremented when creating MMVars based on existing names. The result
was previously that the names are appeared identically would not
compare equal, due to this additional field.

To resolve this problem, contextual name generation is added to the
abstraction functions that produce contexts that are used for
indexing. Internally, abstraction tracks whether a name originated as
a free variable or as an MMVar. Names that originate as free variables
must not be renumbered, because the user chose this name. Names that
originate as MMVars are subject to renumbering.