parser/index: careful treatment of implicit variables
In Harpoon, one could access implicit variables, which significantly
complicates the translation as ordinary Beluga programs can't. More
precisely, it makes translation impossible unless the syntax of Beluga
is somehow extended with implicit abstractions and
applications. Although adding these to Beluga is possible, it is a lot
of work, so it seem preferable to prevent Harpoon from accessing
implicits as well.

To do this, we need to store whether each variable is implicit or not
in the CVar.t context used for indexing. Then, during indexing, we can
emit an appropriate UnboundName error if the user tries to access an
implicit variable. This resolves the issue with the Harpoon
interactive prompt accessing implicits.

However, there is still a problem with Harpoon proof scripts, since
they list the full context at every hypothetical. These contexts now
also need to specify whether each variable is implicit. To do this, I
adjusted the external syntax so the meta-context parsed in a Harpoon
hypothetical can have blank variable names, e.g.
    _ : |- tm
specifies an implicit metavariable of type ( |- tm).
We use the underscore because it is not a valid identifier name.
This is extended to parameter and substitution variables with #_ and
$_ respectively.

Type reconstruction verifies when elaborating a hypothetical that the
plicities of all variables in the given context match the
reconstructed plicities. This is currently buggy because the proof
script that is printed comes from a syntax in which some metavariables
may be marked as inductive, and inducitivity annotations overwrite
plicity annotations. It has now become necessary to separate
inductivity and plicity annotations on metavariables. But this will
have to wait for another patch.